+++ src/pipecat/serializers/telnyx.py
@@ -155,6 +155,10 @@ class TelnyxFrameSerializer(FrameSerializer):
             else:
                 raise ValueError(f"Unsupported encoding: {self._params.inbound_encoding}")
 
+            if serialized_data is None or len(serialized_data) == 0:
+                # Ignoring in case we don't have audio
+                return None
+
             payload = base64.b64encode(serialized_data).decode("utf-8")
             answer = {
                 "event": "media",
@@ -261,6 +265,10 @@ class TelnyxFrameSerializer(FrameSerializer):
                 )
             else:
                 raise ValueError(f"Unsupported encoding: {self._params.outbound_encoding}")
+
+            if deserialized_data is None or len(deserialized_data) == 0:
+                # Ignoring in case we don't have audio
+                return None
 
             audio_frame = InputAudioRawFrame(
                 audio=deserialized_data, num_channels=1, sample_rate=self._sample_rate